+++ CHANGELOG.md
@@ -8,7 +8,8 @@ setBetterPlayerGlobalKey.
 * Updated ExoPlayer version.
 * Added pipStart and pipStop events.
 * [BREAKING_CHANGE] Removed skipsTimeInMilliseconds. Added forwardSkipTimeInMilliseconds and backwardSkipTimeInMilliseconds.
-
+* Updated notification service in android example.
+* Fixed event play/pause event not triggered when controlling video with PiP or remote notification.
 
 ## 0.0.44
 * Added placeholder until play example